Move shared ESLint config under packages
