+++ eslint.config.mjs
@@ -4,12 +4,16 @@ import {
   createNodeJsConfig,
   createNodeTsConfig,
   ignoreConfig,
-} from './eslint/shared.mjs';
+} from './packages/eslint-config/shared.mjs';
 
 export default [
   ignoreConfig,
   createBrowserJsConfig(['frontend/**/*.{js,mjs,cjs}']),
-  createNodeJsConfig(['*.{js,mjs,cjs}', '**/*.config.{js,mjs,cjs}', 'eslint/**/*.mjs']),
+  createNodeJsConfig([
+    '*.{js,mjs,cjs}',
+    '**/*.config.{js,mjs,cjs}',
+    'packages/eslint-config/**/*.mjs',
+  ]),
   createNodeTsConfig(['api/**/*.{ts,tsx}', 'iac/**/*.{ts,tsx}', 'packages/**/*.{ts,tsx}']),
   eslintConfigPrettier,
 ];
