+++ packages/eslint-config/shared.mjs
