+++ packages/shared-types/eslint.config.mjs
@@ -1,5 +1,5 @@
 import eslintConfigPrettier from 'eslint-config-prettier';
-import { createNodeJsConfig, createNodeTsConfig, ignoreConfig } from '../../eslint/shared.mjs';
+import { createNodeJsConfig, createNodeTsConfig, ignoreConfig } from '../eslint-config/shared.mjs';
 
 export default [
   ignoreConfig,
